requirements.txt
Django==3.2.1
djangorestframework==3.12.4
drf-spectacular==0.15.1
